// Client/lajemacarts/src/app/features/dashboard/domain/models/enums/enums.ts
export type ArtworkStatus = 'Draft' | 'InStock' | 'Sold' | 'Archived';


export type WeightCategory = 'LessThan1kg' | 'Between1And5kg' | 'MoreThan5kg';


export type ArtworkType =
  | 'Board'
  | 'Lamp'
  | 'Vase'
  | 'Sculpture'
  | 'PedestalTable'
  | 'IncenseFountain'
  | 'PotPlate'
  | 'Table';

export type DimensionUnit = 'cm' | 'in';